refactoring and renaming the class to match the functionality and usage.
