+++ testdata-generator-common/src/main/java/io/openepcis/testdata/generator/identifier/instances/GenerateGRAI.java
@@ -19,7 +19,7 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
 import io.openepcis.testdata.generator.constants.IdentifierVocabularyType;
 import io.openepcis.testdata.generator.constants.TestDataGeneratorException;
 import io.openepcis.testdata.generator.format.CompanyPrefixFormatter;
-import io.openepcis.testdata.generator.format.RandomMersenneValueGenerator;
+import io.openepcis.testdata.generator.identifier.util.RandomSerialNumberGenerator;
 import io.openepcis.testdata.generator.identifier.util.SerialTypeChecker;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import jakarta.validation.constraints.NotNull;
@@ -99,7 +99,7 @@ public class GenerateGRAI extends GenerateEPC {
   private void generateRandomIdentifiers(final List<String> formattedGRAI, final String prefix, final String modifiedGRAI, final String delimiter, final Long seed, final Integer count) {
     randomMinLength = Math.max(1, Math.min(16, randomMinLength));
     randomMaxLength = Math.max(1, Math.min(16, randomMaxLength));
-    final List<String> randomSerialNumbers = RandomMersenneValueGenerator.getInstance(seed).randomGenerator(randomType, randomMinLength, randomMaxLength, count);
+    final List<String> randomSerialNumbers = RandomSerialNumberGenerator.getInstance(seed).randomGenerator(randomType, randomMinLength, randomMaxLength, count);
 
     for (var randomID : randomSerialNumbers) {
       formattedGRAI.add(prefix + modifiedGRAI + delimiter + randomID);
