+++ testdata-generator-common/src/main/java/io/openepcis/testdata/generator/identifier/instances/GenerateGSRN.java
@@ -19,7 +19,7 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
 import io.openepcis.testdata.generator.constants.IdentifierVocabularyType;
 import io.openepcis.testdata.generator.constants.RandomizationType;
 import io.openepcis.testdata.generator.constants.TestDataGeneratorException;
-import io.openepcis.testdata.generator.format.RandomMersenneValueGenerator;
+import io.openepcis.testdata.generator.identifier.util.RandomSerialNumberGenerator;
 import io.openepcis.testdata.generator.identifier.util.SerialTypeChecker;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import lombok.Setter;
@@ -72,7 +72,7 @@ public class GenerateGSRN extends GenerateEPCType2 {
       } else if (SerialTypeChecker.isRandomType(serialType, count)) {
         // Generate RANDOM identifiers in URN/WEBURI format based on count
         final int requiredLength = maxAppendLength - gcp.length();
-        final List<String> randomSerialNumbers = RandomMersenneValueGenerator.getInstance(seed).randomGenerator(RandomizationType.NUMERIC, requiredLength, requiredLength, count);
+        final List<String> randomSerialNumbers = RandomSerialNumberGenerator.getInstance(seed).randomGenerator(RandomizationType.NUMERIC, requiredLength, requiredLength, count);
 
         for (var randomID : randomSerialNumbers) {
           formattedGSRN.add(prefix + gcp + delimiter + randomID);
